Visual changes to class info cells
Moved the teacher to under the class name, and made the class name bold to stand out. Removed notes from initial view, need to create a way to signify to the user that it is possible to click it to get the full note.

Current on the right-hand side of the cell is a > arrow, its y-coor needs to change but Apple seemingly does not allow it.